Permite agregar gastos aprobados a actividad

+++ db/migrate/20230516135641_agrega_presupuesto_aml.rb
@@ -1,6 +1,6 @@
 class AgregaPresupuestoAml < ActiveRecord::Migration[7.0]
   def change
-    add_column :cor1440_gen_actividadpf, :presupuesto, :integer, 
+    add_column :cor1440_gen_actividadpf, :presupuesto, :decimal, 
       null: false, default:0 
   end
 end
